backend/api/pyproject.toml
[project]
name = "loopweave-api"
version = "0.1.0"
description = "LoopWeave API Service"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "google-cloud-bigquery==3.13.0",
    "google-cloud-firestore==2.13.0",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "httpx==0.25.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

